Non-malicious symbols must be reverted to the original text.
As a part of sanitizing the incoming request parameters the strings are converted to the respective symbol. If the symbol is non-malicious then it must be reverted to the original text. This is happening when text is having '&' followed by characters which may get evaluated to symbols. for e.g. &Mu is evaluated to the Greek symbol µ.